Added PointLights and DirectionalLights to ForwardRenderer

+++ LearningEngine/src/Graphics/Renderer/API/ShaderBuffer.cpp
@@ -3,7 +3,7 @@
 #include "RendererAPI.h"
 #include "Platform/OpenGL/OGLShaderBuffer.h"
 
-ShaderBuffer* ShaderBuffer::Create( uint32_t size, ShaderType type)
+ShaderBuffer* ShaderBuffer::Create(__int64 size, ShaderType type)
 {
     switch (RendererAPI::GetAPI())
     {
